Moved cascading-scan tests from tests/integration to cascading scan hook folder
This is done so these tests would run when "make integration tests" is called
in the cascading-scan hook folder
This goes hand in hand with issue #890
